api/requirements.txt
appdirs
asgiref
attrs
blinker
certifi
charset-normalizer
click
colorama
coreapi
coreschema
Django
django-axes
django-cors-headers
django-csp
django-debug-toolbar
django-extra-checks
django-health-check
django-http-referrer-policy
django-ipware
django-migration-linter
django-permissions-policy
django-querycount
django-split-settings
django-stubs-ext
django-test-migrations
djangorestframework
djangorestframework-simplejwt
drf-spectacular
drf-spectacular-sidecar
drf-yasg
idna
inflection
itypes
Jinja2
jsonschema
MarkupSafe
nplusone
numpy
packaging
Pillow
PyJWT
pyparsing
pyrsistent
python-decouple
pytz
PyYAML
requests
ruamel.yaml
ruamel.yaml.clib
six
sqlparse
structlog
toml
typing_extensions
tzdata
uritemplate
urllib3
